requirements.txt
PyQt6
ffmpeg-python
gTTS
pysubs2
APScheduler
requests
Pillow
